+++ src/app/(basic)/contact/page.tsx
@@ -1,0 +1,11 @@
+import { CreateContactForm } from '@/src/app/(basic)/contact/_component/CreateContactForm/CreateContactForm'
+
+const Page = () => {
+  return (
+    <>
+      <CreateContactForm />
+    </>
+  )
+}
+
+export default Page
